Add Specimen Skeleton char categories
This code is copied over from the Specimen Skeleton project and
only lightly patched so work in the WF engine. This code works,
but it kinda dumb. There's a refactor issue in the Specimen
Skeleton project, if that is done we should improve the code
here as well.

Another small point of concern is that the GlyphData is over 5MB
large! This is a pretty hefty download, so we need to be smart
about loading this asset.